Prepare version 1

+++ src/bitmap/pixel_data.rs
@@ -162,10 +162,8 @@ impl IndexMut<usize> for PixelData {
 #[cfg(debug_assertions)]
 impl std::fmt::Display for PixelData {
     fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        for p in 0..self.pixels.len()
-        // 0..std::cmp::min(5, self.pixels.len()) as usize
-        {
-            write!(f, "{}: {}\n", p, self.pixels[p as usize]).unwrap();
+        for p in 0..self.pixels.len() {
+            write!(f, "{}: {}\n", p, self.pixels[p]).unwrap();
         }
         write!(f, "")
     }
